后台:1)完成用户接口2)完成歌曲接口

+++ models/loginLog.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type LoginRecord struct {
-	ID          uint `gorm:"primaryKey"`
-	UserID      uint
+	ID          uint64 `gorm:"primaryKey"`
+	UserID      uint64
 	LastLoginIP string
 	LoginTime   *time.Time `gorm:"autoCreateTime"`
 }
